Issue number:  #25 - "Ability to delete a log/build" Contrubuted by:  Vladimir Sizikov Reviewed by:   Kohsuke
So, a new button, similar to "Keep this log" was added: "Delete this build".

The new button only enabled if the build is not marked with "keep this
log forever" - there is no need to delete a build if the user explicitly
stated that this build must be preserved.

When user click on the "Delete" button, he is redirected to
confirmation page, similar to what is currently available for "Delete
Project".

We once again double-check that it's OK to delete the build, and there
are no other projects that depend on this build and wish it to stay.

If we see that the build should not be deleted, a warning shown to
user.

If it's OK to delete the build and user clicks "Yes" button, we delete
it.


git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@1067 71c3de6d-444a-0410-be80-ed276b4c234a