[FIXED HUDSON-2373] Modified to work with Maven 2.1. Sorry for a long delay, folks. This change will be in 1.305.
We basically need an ability to feed different components.xml depending on Maven's version (which will be read by DefaultComponentDiscoverer in Plexus.) The implementation in this fix is hacky, and based on placing one more jar in front of the classpath when run on Maven 2.1, while maven-interceptor still contains the default components.xml for 2.0.x.

Ideally 2.0.x components.xml should move to its own jar and so that we can switch between them, as opposed to pass in extra jar only when it's Maven 2.1.

We'll see how Maven 3.0 works with this regard.

git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@18036 71c3de6d-444a-0410-be80-ed276b4c234a