fixed a race condition in a test.
mindless discovered that m.scheduleBuild2(0).get(0) has a cascading effect on module builds, which can cause later m.getModule(...).scheduleBuild2(0) to return null, as it may be already in the queue.

I initially wanted to fix this by making Queue.schedule() returns the existing item if it's already in the queue, but this turns out to be inpractical given that the signature is WaitingItem and not Item.

git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@22991 71c3de6d-444a-0410-be80-ed276b4c234a