Honor $JENKINS_HOME.
Honor $JENKINS_HOME (just like we used to do $HUDSON_HOME)
We still fall back to $HUDSON_HOME every step of the way
for maximum backward compatibility.

I wasn't initially planning to do this for the first version,
but working on native packages made me think we should deliver this.