Revert "merge changelog.html fixes from kk"
Note that the above description is incorrect as to what the merge was about.
It is the message that I intended to make. It appears that I accidentaly merged multiple fixes waiting to be reviewed on my private fork.

This reverts commit dce0d2aff0a440d7405b9ba10fafa2f3fe98a27a, reversing
changes made to 63f67f30bdcfd062bccc85d245109b34785a195b.