Fix dependency on Java2 runtime for both Debian and Ubuntu
In Debian, a dependency on 'java-runtime' virtual package may be
satisfied by pre-Java2 JVM [1] which is wrong. According to Debian
policy for Java [2], virtual machines that contain a runtime
conforming to the Java 2 specification should provide java2-runtime
[3] and Java programs must depend on the needed runtime environment
(java1-runtime and/or java2-runtime) [4].

As Jenkins' target is 1.5+, it's better to depend on Java2 JVM [5]
explicitly.

This should also work in Ubuntu as Sun Java5 [6],
Sun Java6 [7] and OpenJDK [8] packages provide 'java2-runtime' too.

[1] http://packages.debian.org/sid/java-runtime
[2] http://www.debian.org/doc/packaging-manuals/java-policy/
[3] http://www.debian.org/doc/packaging-manuals/java-policy/c43.html#policy-vm
[4] http://www.debian.org/doc/packaging-manuals/java-policy/x85.html
[5] http://packages.debian.org/sid/java2-runtime
[6] http://packages.ubuntu.com/hardy/sun-java5-jre
[7] http://packages.ubuntu.com/hardy/sun-java6-jre
[8] http://packages.ubuntu.com/maverick/openjdk-6-jre

Signed-off-by: Mykola Nikishov <mn@mn.com.ua>