[FIXED JENKINS-8647] Many messages about RecordReaper IllegalArgumentException
I did a checkout for our internal fork of JmDNS [1] and the
upstream version [2] to compare what the differences were..
I found none. There do not appear to be any local edits to
our fork so I can not figure out why we did the fork to
begin with.

Perhaps some older version had locally patched bugs which
have since been resolved? Anyhow, given that JENKINS-8647
is causing lots of people lots of grief, and it is enabled
by default, I went ahead and bumped to jmdns-3.4.0 upstream.

I think reduction of forks for library projects is generally
good, and especially when we have no practical changes made
to our version.

[1] https://github.com/jenkinsci/jmdns
[2] https://jmdns.svn.sourceforge.net/svnroot/jmdns/tags/JmDNS_3_2_1/