[FIXED JENKINS-8579] defined a mechanism to replace some of the key UI text.
I wasn't initially planning to do this, but it turns out more people needed this feature than I thought.

I left the text as is, but defined a mechanism to override this in the plugin, in a possibly context sensitive way. For example, a plugin can replace this text on a job-by-job basis, or by looking at some property/builder/etc.