Allow maven builds to (opionally) make use of the token-macro-plugin.
It is useful to be able to pass parameters in the form of
-Doption=${THING}
to the maven builds.

The new token-macro-plugin stuff from KK is ideal for this.

To get the actual parameters through to the launch commandline requires
some adjustments to the Launcher API, because it is not aware of the
build itself, which is neccessary to do the options expansion. So instead
pass these through to the launcher itself.

If the token-macro-plugin is not installed, the behaviour remains as before.

Signed-off-by: Nigel Magnay <nigel.magnay@gmail.com>