[FIXED JENKINS-8748]: openSUSE: required JDK
For openSUSE it doesn't make sense to require a specific flavor of JRE
for Jenkins. Even if the RPM explicitly requires the Oracle JDK, thus
causing it to be installed together with Jenkins, some previously installed
JRE may still be used to run Jenkins, due to the update-alternatives(8) setting.
The JRE to use is a sysadmin issue, and not usually part of the RPM spec.

I have not been able to reproduce the scenario described by David.
Installing Jenkins on fresh installs of openSUSE since v11.1 pulls in openJDK
1.6 with it, with which Jenkins runs just fine. Even on Live CDs.