Integrated a module that retains an RSA key pair as an identity.
This is a proof of concept to core modulizarion (by binary integration.)
A module is built like a plugin but it's packaged to the war as a
WEB-INF/lib/*.jar, not as a plugin.

Compared to plugins, jars integrated in this way do not double-package
dependency jars, and they are not visible to the users. I plan to try a
few of these and if it's useful, this could be an useful mechanism for
modularizing core.