Move hudson.model.Jenkins to be jenkins.model.Jenkins - part I
Because some parts of the API rely on package protected accesses between
various parts of code in hudson.model.*, create a new base class (AbstractCIBase)
which contains just those parts, and create utility functions for Jenkins
to call. This means minimal disruption to other parts of the system.

Minor change to deprecated function in Hudson to use an available API,
rather than access to a protected variabl.

Signed-off-by: Nigel Magnay <nigel.magnay@gmail.com>