[FIXED JENKINS-9915] Make captcha support optional; remove LGPL jcaptcha
This change is largely based on changes made by Winston Prakash for Hudson.
The main details about his changes can be found here:

http://issues.hudson-ci.org/browse/HUDSON-8844

I deviated a bit from his changes where it was appropriate (e.g. the
description for allow-signup.html was just plain wrong). Also, there
was a lot of collateral damage from formatting in his changeset, which
I have factored out.

I tested the basic functionality of this change and it seems to be OK.