Switch to a fully published packaging interface
The main benefit of this is that the PackageMaker
program doesn't have to be used anymore. The
problem with Packagemaker is that the file format
is supposed to be opaque, but we were fiddling
with it to introduce automated changes. This
became an issue as PackageMaker updates the file
format causing a new diff every time the file is
opened.

By using a commented XML file, future changes are
easier to integrate, and newcomers can get a
handle on this facet of the project faster.