Activate the login Keychain on OSX
On OSX, especially for iOS and App-Store Development, access to the login keychain is required to the certificate based code signing with Xcode.
The following patch enables access to the Jenkins account login keychain that should contain a valid copy of the developer certificates. Besides enabling access to the keychain, users must grant access to these certificates to Xcode using the keychain application.