Report remote address of the client, not the forwarding proxy in token authorization.
When running Jenkins behind a proxy (nginx, apache), the remote address
reported in the build cause will always be the one of the proxy,
normally localhost. While there is an alternative solution when using
tomcat (RemoteIpValve, see
http://kisdigital.wordpress.com/2011/08/19/retain-remote-address-when-proxying-railo-with-nginx/)
winstone seems to have no such valve. Therefore take the content of the
X-Forwarded-For header when it is available.