[FIXED JENKINS-11089] Added logic behind "avoid fallback JDK" option in axes configuration UI
While building matrix projects, the logic checks for an empty JAVA_HOME
environment variable to decide whether to skip the run in case the
default fallback JDK is going to be used (JAVA_HOME is empty or equal
to "") or to proceed with the build and anything is fine. Uses the
JDK_NOT_AVAILABLE build result and therefore displays as grey ball
without changing the overall build result. JENKINS-11089