Log with a specific user when controlling Windows slave
Add an additional option in "Let Jenkins control this Windows slave as a
Windows service" to specify a user under which the service will be run.

This allows to use a different user than "Local System" to register a
windows service.

Added as well a few fr translations for strings in this config page.