Revert "Revert "Don't provision more slaves when some are coming online.""
This reverts commit 47c520b13bfb08d0fcc9c01a397a872ad5405a24.

I see it clearly now. I was bit concerned about double-counting,
because one expects that totalSnapshot (total # of executors)
includes idle executors. But this is not the case, and in any way
this code should be read as (x==0 && y==0 && z==0) not (x+y+z==0)
conceptually. So reverting my revert.