follow up fix to c51114a1970738dcc7d178c486fda03a7ada8755.
This still is slightly more robust than the original in the sense that it's not affected by the presence/absence of the trailing '/' in itUrl.

----------------------
Comment from @ohtake:

getContextPath() will return "" if Jenkins is served at ROOT.

Assume that getContextPath() => "", itUrl => "iturl" and urlName => "urlName",
value returned was "/iturl/urlName", but is "iturl/urlName" (no slash at 0).