[JENKINS-13190] Add ACLPermissionOverride extension point
Modifies the base ACL class to first check with ACLPermissionOverride extension point
implementations as to whether the permission should be granted.

If none of the ACLPermissionOverride instances approve the permission the ACL's permission
checking logic will be used instead.

This override provides a mechanism to grant more permissions than the ACL being overridden provides.