osx: Self-extracting command-line installer
Re-designed the installer to not depend on Apple Installer.

Now the installer runs completely on command-line, making it
possible to install/upgrade over ssh.

The installer is interactive. It allows the user to customize
several aspects of the installation, currently:

* Should Jenkins be started at boot?
* Which user should be used to run Jenkins?

The self-extracting archive is provided by makeself.
Although the makeself license is GPL, it explicitly says the
contents of the archive do not fall under GPL.