Reworked the dynamically modifiable Axis.
I thought about this and decided that the best way to incorporate this
expansion of the abstraction is to think of it as a re-configuration of
project right before the build.

This resolves my concern with the previous change of calling expensive
rebuildConfigurations() from every getActiveConfigurations() call (and
getter shouldn't have a side-effect, right?)

Also, in this way, Axis gets a reference to MatrixBuildExecution, so it
makes it more explicit that the Axis can update values based on the
current build.