[FIXED JENKINS-12318]
Preserve in-progress builds when reloading a job.
Because in-flight builds tend to update the state a lot, with this
change we refrain from reloading those builds from the disk.

This should be acceptable since we are primarily reloading a job, and
reloading of builds are secondary.
(cherry picked from commit 9fbd6d3a482643f4080fbb793c67323e8464b11c)

Conflicts:

	changelog.html