Allow projects to be blocked while downstream projects are also blocked.
Adds an extra option to the configuration to allow projects to be blocked when downstream jobs are also blocked.
This is critical if a chain of jobs must run once for an entire chain
and there is a diamond pattern at any point in the chain.
as the job after the parallel will be blocked if it is waiting for the upstream jobs to finish,
and the start job may then run again as the joining job is not in a state to run.
[FIXED JENKINS-14918]