Enabled transparent compression support.
This only works if the "transparent GZIP support" patch has been
applied against Stapler. Otherwise, this patch will not compile
as the new "LargeText" constructor will not be found.

Additionally, the console.jelly was modified to make use of the
stream instead of the raw file, which is necessary to get the
correct uncompressed size of the file for skipping bytes.

[JENKINS-2551]
[JENKINS-10400]
[JENKINS-13655]

Signed-off-by: Martin Schroeder <martin.h.schroeder@intel.com>