Create a marker interface for actions to beable to pass matrix configurations
create a Interface that Actions can implement to indicat that they
should be passed from
the matrix parent to the indiviual configurations.
This means that the SVN Parameters can be passed to them.
Partial fix for [JENKINS-14619] also needs changes in SVN SCM plugin to
fix completely.

Developed by cjo9900