[JENKINS-15465]
If a build record gets GCed, a load is attempted, and that fails, then RunMap can return null. Not sure if this alone would explain the entirety of JENKINS-15465, but this is a related fix.
(cherry picked from commit ca44a90e27263d390ff5414b25ec4e7b5c79d687)