if there's only one installation, better to resolve to that one.
After all, that's the installation the user is most likely going to choose, and if there's only one installation.

Plus when there's only one installation, there's no drop-down to select the Maven installation in the config UI, so the user is asked to do a rather unnatural "open config page" > "just click save" workflow.