Make sure that change log annotation cannot result into exception
Although javadoc for ChangeLogAnnotator#annotate() states that any exception should be caught in this method and it should return normally, some implementation don't respect it or e.g. in case of missing some dependecy (resulting into NoClassDefFoundError) plugin itself can hardly ensure that no exption is thrown. We should ensure it on the core level.