Add finalizePlannedNode() method to Cloud objects.
For plugins that implement Cloud objects, it can be important for them
to be notified when a PlannedNode instance is about to be destroyed (for
example, this can be used to track the number of active and planned instances
in the cloud reliably). This patch adds an optional method to the Cloud
class that is called by the NodeProvisioner, if the Cloud has requested it,
to indicate that the NodeProvisioner is finished with a PlannedNode object.