Fixing a test regression.
See 010b47d7307a11f0750fc01a00d99b96bbab9234 in stapler/stapler for the
actual fix.

I first discovered this on my local desktop yesterday as I was verifying
the fix for json-lib 2.4. From what I can tell, this has been broken all
along and it was working only because Facets are discovered in the lucky
order; JellyFacet got a chance to get at it before GroovyFacet did.

When I rebuilt from clean the problem stopped appearing, so I assumed
that there are some factors that cause JellyFacet to get loaded before
GroovyFacet (maybe the order in which Maven iterates dependencies or
something like that.)

But I'm no longer so sure now that I know J-on-J saw this problem at
about the same time. That's just too many coincidences.

But let's see if this would make J-on-J happy.