Allowing the object returned from ArtifactManager to be bound under the URL space of build.
DirectoryBrowserSupport is a bit of weirdo, in that its a very intelligent HttpResponse object that creates a whole hierarchy. But most of the time, complex structures are created by a series of POJOs referencing each other. Defining a signature to return HttpResponse prevents this common design pattern.

In contrast, as I wrote in the revised javadoc, one can force a certain HttpResponse during the object traversal by throwing an HttpResponse.

So this new signature can accommodate both cases, so I recommend this over the original signature