Deprecated jobBaseUrl from projectView.jelly.
The reason is that we always determine the job url using
Functions#getRelativeLinkTo, which keeps the view navigation info if it
exists, and can produce an absolute url if required by the context. So
it is no longer needed for the caller to worry about jobBaseUrl.

By default, names are interpreted in relatively to the current item
group, so Label and Computer views no longer need to set useFullName to
true to have correct behaviour.