Don't use target/classes to unpack maven installs.
This causes the jar plugin to see the results of any maven installations used during unit testing and significantly bloats the
resulting jar/hpi/jpi size.

This moves the location of maven installs to a separate folder under target so that it should not conflict with any maven plugins.

The use of "buildDirectory" from a property is questionable - but left as-is incase anyone was relying on this propertyname.