Add classes and ID's for common elements
This should aid theme developers in developing plugins for Jenkins by adding
CSS classes and ID's to the dom. Developers can then use these to more
specifically target elements that they want.

For example, the "Console Output" was a raw <pre> tag embedded in a <table>.
Styling it required CSS like so:

    #main-panel > pre {

Sadly this will apply to any <pre> elements nested directly under the
"#main-panel" ID; there's no way to get more specific to the console output.

With this change you can style the console output by writing:

    .console-output {

Which is much more direct (and more efficient for browsers to render).

A similar treatment is applied to a number of elements in the jenkins UI.