Fix JENKINS-19173
Added Action.ContextMenuVisibility annotation and a method to Functions to determine if the annotation is present (if it is not, then the Action will be visible) and check if the Action is marked as not visible. Updated actions.jelly to use this method and set the contextMenu attribute. Also, added a test to test the annotation.