[FIXED JENKINS-21729] Fix IE8 NPE with respect to sortability
IE8 (and I assume other browsers that use a doctype pre-html5) interpret the span outside of the link, as in HTML4.01 you needed the closing tag.
http://www.w3.org/TR/REC-html40/struct/global.html#h-7.5.4

The code was causing a NPE in IE8 and users could not sort the columns.

>> cell.innerHTML = '<a href="#" class="sortheader">'+this.getInnerText(cell)+'<span class="sortarrow" /></a>';

>> cell.innerHTML ;
"<A class=sortheader href="#"> S<SPAN class=sortarrow></A></SPAN>"

>> cell.innerHTML = '<a href="#" class="sortheader">'+this.getInnerText(cell)+'<span class="sortarrow"></span></a>'

>> cell.innerHTML ;
"<A class=sortheader href="#"> S<SPAN class=sortarrow></SPAN></A>"