Move invalid body ID to body class and data
IDs cannot contain spaces. Instead, let's move jenkins-${h.version} to the `class` attribute, which is space delimited. This way it can still be used for styling if needed (using `body.jenkins-x.y`). If this is not used for anything other than knowing which version of Jenkins is running, use `<body id="jenkins" class="yui-skin-sam" data-version="jenkins-${h.version}">` -- IDs and Classes should not be used to store data.