[FIX JENKINS-17114] Distinguish 'nodes offline' from 'no nodes'
Jenkins currently does not distinguish between 'All nodes for a
label are offline' and 'There are no nodes for a label'. This is
exacerbated by the current wording:

    All nodes of label 'foo' are offline

Instead of just changing the wording to something not seeming to
imply existence of such nodes, add a second wording that clearly
states that there are no such nodes.