Javadoc: Noting changes in Executor availability check approaches in 1.536
This change is a follow-up to Executor::isAlive() issues in #1229.

Since 1.536, {@link Executor}s start threads on-demand.
The entire logic should use Executor::isActive() instead of Executor::isAlive() in order to check if the {@link Executor} it ready to take tasks.

Since the change breaks previous executor availability flows, it is required to document changes somehow.

Signed-off-by: Oleg Nenashev <o.v.nenashev@gmail.com>