Switch font stack to Helvetica, Arial
While Verdana and Georgia are widely available on user machines, they have
become a rare sight in web interfaces today, and are more commonly associated
with old websites. The list of fonts that are on most users computers is small,
and few of those are acceptable to use as a general-purpose web font for
Jenkins. This leaves Helvetica and its close cousin Arial; Helvetica looks
slightly nicer but its cousin is deployed on many more machines, so it provides
a good fallback.

Change the default font color #000 -> #333. This provides a slightly nicer
contrast with pure white, without hurting accessibility.

Change "Jenkins" title to a link, and the font from Georgia to Helvetica. Text
plays better with SEO and reduces the bytes needed to be transferred over the
wire.