Convert Jenkins background image to grayscale
The idea here is to reduce contention for the user's eye in the display.
Color is generally an indication of important UI elements, and in this case
using color in the background takes away from the important elements in the
foreground (generally the job menu or the build history sidebar).

Changing the Jenkins logo to grayscale should help with this.

In addition, this moves some of the styles into the main Jenkins stylesheet and
out of the HTML.