Switch underline/non underline on sidebar
The underline adds slightly too much visual weight to the links in question.
This changes the links to not have an underline by default, while still adding
the underline when the link is hovered.

Normally accessibility requirements demand that you add an underline for all
clickable text. However, in this case the positioning of the elements, their
changed color, the presence of an icon next to the links, and the link hover
effect combine to give these elements the appearance of clickability.